:bug: fix function names
